+++ app/src/main/java/com/example/ehasibu/utils/api/APIConsumer.kt
@@ -106,7 +106,7 @@ interface APIConsumer {
     suspend fun updateCustomer(@Body customer: UpdateCustomerRequest): Response<ApiResponse<CustomerResponse>>
 
     @DELETE("customers/{id}")
-    suspend fun deleteCustomer(@Query("customerId") customerId: Int): Response<ApiResponse<CustomerResponse>>
+    suspend fun deleteCustomer(@Query("customerId") customerId: Long): Response<ApiResponse<CustomerResponse>>
 
 
     //Purchases
